+++ src/Better-comment.py
@@ -1,0 +1,26 @@
+"""
+ _______       _            _     _          ______        _                 _ 
+(_______)     (_)       _  (_)   | |        (____  \      (_)               | |
+ _______  ____ _  ___ _| |_ _  __| |_____    ____)  ) ____ _ _____ ____   __| |
+|  ___  |/ ___) |/___|_   _) |/ _  | ___ |  |  __  ( / ___) (____ |  _ \ / _  |
+| |   | | |   | |___ | | |_| ( (_| | ____|  | |__)  ) |   | / ___ | | | ( (_| |
+|_|   |_|_|   |_(___/   \__)_|\____|_____)  |______/|_|   |_\_____|_| |_|\____|
+    
+Auteur: Frank SAURET(frank.sauret.prof@gmail.com) 
+t.py(Ɔ) 2022
+Description : Saisissez la description puis « Tab »
+Créé le :  vendredi 7 janvier 2022, 16:39:14 
+Dernière modification : vendredi 7 janvier 2022, 17:06:06
+"""
+
+# Commentaire non colorisé
+# * Commentaire important
+# ! Faire attention
+# ? Plutôt pour de l'aide
+# todo : Les trucs à faire
+# § Ca c'est super important
+# $ Et puis ça aussi
+# 1- Pour structurer le document
+# 2- Celui ci aussi 
+# 3- Et un petit dernier
+